terse1/daacs_doc_system/requirements.txt
# DAACS v3.0 Document System Dependencies

# Core
langgraph>=0.0.30
langchain-core>=0.1.0
langchain-openai>=0.0.5

# Optional: For different LLM providers
# langchain-anthropic>=0.1.0
# langchain-google-genai>=0.0.5

# Utilities
python-dotenv>=1.0.0
